pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jibon-sathu-backend"
version = "0.1.0"
description = "Zone A cloud API — road marking state, SOS, supply sync"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.29.0",
    "alembic>=1.13.3",
    "pydantic-settings>=2.6.0",
    "greenlet>=3.1.0",
    "bcrypt>=4.2.0",
    "slowapi>=0.1.9",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "grpcio>=1.62.0",
    "packaging>=24.0",
    "protobuf>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "grpcio-tools>=1.62.0",
    "psycopg2-binary>=2.9.9",
    "onnxruntime>=1.19.0",
    "onnx>=1.16.0",
    "scikit-learn>=1.4.0",
    "skl2onnx>=1.17.0",
    "numpy>=1.26.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
